Fix controller security issues and user password handling

+++ src/Controller/Admin/UserController.php
@@ -9,12 +9,15 @@ use Doctrine\ORM\EntityManagerInterface;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
+use Symfony\Component\PasswordHasher\Hasher\UserPasswordHasherInterface;
 use Symfony\Component\Routing\Attribute\Route;
 use Symfony\Component\Security\Http\Attribute\IsGranted;
 
 #[IsGranted('ROLE_ADMIN')]
 class UserController extends AbstractController
 {
+    public function __construct(private UserPasswordHasherInterface $passwordHasher) {}
+
     #[Route('/admin/users', name: 'admin_user_index')]
     public function index(UserRepository $repo): Response
     {
@@ -27,11 +30,12 @@ class UserController extends AbstractController
     public function create(Request $request, EntityManagerInterface $em): Response
     {
         $user = new User();
-        $form = $this->createForm(UserType::class, $user);
+        $form = $this->createForm(UserType::class, $user, ['require_password' => true]);
         $form->handleRequest($request);
 
         if ($form->isSubmitted() && $form->isValid()) {
-            $user->setStatus(1);
+            $plainPassword = $form->get('plainPassword')->getData();
+            $user->setPassword($this->passwordHasher->hashPassword($user, $plainPassword));
             $em->persist($user);
             $em->flush();
 
@@ -68,10 +72,17 @@ class UserController extends AbstractController
     #[Route('/admin/users/{id}/delete', name: 'admin_user_delete', methods: ['POST'])]
     public function delete(User $user, Request $request, EntityManagerInterface $em): Response
     {
+        if ($this->getUser() === $user) {
+            $this->addFlash('error', 'You cannot delete your own account.');
+            return $this->redirectToRoute('admin_user_index');
+        }
+
         if ($this->isCsrfTokenValid('delete_user_' . $user->getId(), $request->request->get('_token'))) {
             $em->remove($user);
             $em->flush();
             $this->addFlash('success', 'User deleted.');
+        } else {
+            $this->addFlash('error', 'Invalid security token. Please try again.');
         }
 
         return $this->redirectToRoute('admin_user_index');
